fix extraneous midi realtime messages on playback (#4822)
* apply skipped ticks to the ramp that was actually running

ParamManagerForTimeline::processCurrentPos accumulates ticksSkipped and clears
it only when the interpolation block runs, which happens at automation nodes -
so it grows to the whole gap between one node and the next. It then called
processCurrentPos() first (which reaches the node and installs a fresh
valueIncrementPerHalfTick for the *upcoming* span) and applied ticksSkipped to
that new increment.

For MIDI expression automation that means a whole inter-node gap's worth of a
few-tick ramp gets applied in one step. On the song from #4821, note row y=56
sits 196 ticks with nothing interpolating, then hits a node whose next segment
is 6 ticks long; the 196-tick step sends per-note pressure to -16 instead of 6,
and the ramp back up emits aftertouch values of -6 and -4.

Tick the collection before processCurrentPos() instead, so the elapsed ticks
advance the span they belong to, and re-check afterwards for whether a node has
just started some interpolation we need to come back every tick for.

* saturate the interpolation increment instead of wrapping it

tickTicks() and tickSamples() scaled valueIncrementPerHalfTick by the elapsed
time in int32 arithmetic. A steep increment - a short segment covering a big
value change - times a large tick count overflows easily, and a wrapped product
drives currentValue hard in the wrong direction: on the #4821 song an increment
of 20971520 over 196 ticks computes 8220835840, which wraps to -369098752.

add_saturate() on the result only guards the addition, not the multiply, so
clamp the product to the int32 range first.

* saturate MIDI expression data bytes into 0-127

Aftertouch and CC values are derived from internal 32-bit parameter values by
shifting, so they can land outside 0-127 - MIDIInstrument passes value32 >> 24
straight through, and callers in MIDIDrum and Sound do the same. Neither
sendSerialMidi() nor setupUSBMessage() masks data bytes, so an out-of-range
value goes onto the wire with bit 7 set and the receiver reads it as a status
byte. Values landing in 0xF8-0xFF get read as spurious System Real-Time
messages, which is what #4821 reports - the Deluge has no way to send 0xFF at
all, so the Reset messages in that log were negative pressure values.

Clamp in MidiEngine rather than at each call site, so paths that don't go
through MIDIInstrument are covered too, and widen the aftertouch value
parameters from uint8_t so a negative saturates to 0 instead of truncating.
sendCC() already clamped the top end; give it the bottom end as well.

+++ src/deluge/io/midi/midi_engine.h
@@ -92,8 +92,10 @@ public:
 	///
 	/// @param bend Bend amount. Only the lower 14 bits are used
 	void sendPitchBend(MIDISource source, int32_t channel, uint16_t bend, int32_t filter);
-	void sendChannelAftertouch(MIDISource source, int32_t channel, uint8_t value, int32_t filter);
-	void sendPolyphonicAftertouch(MIDISource source, int32_t channel, uint8_t value, uint8_t noteCode, int32_t filter);
+	/// @param value Pressure amount. Saturated into 0-127
+	void sendChannelAftertouch(MIDISource source, int32_t channel, int32_t value, int32_t filter);
+	/// @param value Pressure amount. Saturated into 0-127
+	void sendPolyphonicAftertouch(MIDISource source, int32_t channel, int32_t value, uint8_t noteCode, int32_t filter);
 	bool anythingInOutputBuffer();
 	void setupUSBHostReceiveTransfer(int32_t ip, int32_t midiDeviceNum);
 	void flushUSBMIDIOutput();
